Initial ajanottaja server implementation
Implements initial Ajanottaja API with support for starting and stopping
intervals. Integrates with Auth0 via endpoints that is called upon Auth0
register and login events.

Uses http-kit as the web server of choice. Reitit handles routing and
integrates with malli to handle coercion/transformation of input and
output values as well as data validation.

Uses next.jdbc to communicate with Postgres and honeysql to generate
queries.

An initial scaffold for the frontend is provided using figwheel,
reagent, and tailwind for css stuff. Webpack is used to bundle
javascript dependencies.

Signed-off-by: Snorre Magnus Davøen <post@snorre.io>