Add url env and autofill Swagger oauth client id
When running swagger ui it is annoying to have to find the Oauth2 client
id for the authentication to work. This is now fixed by hardcoding the
client id in a new swagger-ui/index.html resource file which replaces
the one provided by metosin's swagger-ui library.

The server config now also expects a new :url variable which tells the
server which scheme + hostname the server is running behind so that
redirect URLs can be constructed programatically.

Signed-off-by: Snorre Magnus Davøen <post@snorre.io>