Remove client code and simplify namespaces
The client will henceforth be a separate project written with Vite.js
and React. The backend and shared namespace parts have been removed. The
API docs will be served from the root path. It is expected that you run
the backend on a separate domain from the frontend, e.g.
api.ajanottaja.app.

Signed-off-by: Snorre Magnus Davøen <post@snorre.io>