Camelcase swagger spec to fix swagger doc issues
The swagger spec would previously report properties in a mix of
kebab-casing and camelCasing. This was caused by the spec being
serialized to JSON by muuntaja which uses a camelCasing plugin to camel
case the map keys in the responses. However non-keys were not being
camelCased and so a mismatch would occur where lists of required
properties would be kebab-cased and thus not properly handled by swagger
tooling.

The server now uses an interceptor to properly camelCase all keywords in
the generated swagger spec to avoid the problem.

Signed-off-by: Snorre Magnus Davøen <post@snorre.io>