Add calendar endpoint with calendar like data
A new endpoint /calendar can be used to fetch a list of dates with
target duration and logged intervals for each date. The endpoint
supports listing calendars in day, week, month, and year steps.
The caller is expected to supply a start date to use as first date in
list.

If no intervals has been logged for a given date, the list of intervals
for that date is empty. If no target duration is set for a given date,
its value will be null.

Signed-off-by: Snorre Magnus Davøen <post@snorre.io>