Rename docker image to ajanottaja/api
Considering the project is currently split into separate api and
frontend projects the docker image name should reflect this. The
frontend is currently built and hosted using Cloudflare Pages, but there
is likely that a Docker image should be made for the frontend to support
other hosting solutions for the frontend as well.

Signed-off-by: Snorre Magnus Davøen <post@snorre.io>