Avvoid running update check within project folder
When running the update action from the same folder as the source code,
it loads the dependencies (including logging dependencies and others)
that the project using polluting the logs the update check bases its
update checks on. Checking out the code to a subfolder and running the
update tool from the outside should hopefully fix this issue.

Signed-off-by: Snorre Magnus Davøen <post@snorre.io>