Fix calendar statistics bug
The calendar statistics endpoint was querying the database in such a way
that target durations were repeated and summed up per interval recorded
for a given day. So working two intervals and having a target of 8 hours
produced a combined 16 hour target instead of the expected 8 hours.

This is now fixed by looking for the minimum value. This is not an
optimal query, but good enough for now. In the future it might be better
to query and aggregate the interval table separately before doing the
outer join and diffing the values.

Signed-off-by: Snorre Magnus Davøen <post@snorre.io>