Cumulative diff endpoint for tracked vs target
Adds new endpoint with a cumulative sum of the difference between the
tracked intervals and targets for each date with at least one interval
or target registered.

The cumulative value can be used to draw a chart showing the development
of the positive or negative difference in tracked interval durations and
registered targets for each date.

Signed-off-by: Snorre Magnus Davøen <post@snorre.io>