Initial auth0 bootstrapping support
Contains a new namespace, ajanottaja.auth0, that can be used to
bootstrap the Auth0 setup. Given a valid configuration it uses the Auth0
managment API to check the current state of actions on Auth0 and then
creates or updates as necessary. Currently it will always rebind the
actions to the triggers. It should first check if the triggers have been
bound before rebinding them.

The logic will currently not be run on startup as this is a slow
process. Instead the Ajanottaja api project should expose some
functionality for running these kinds of operations, preferably on the
command line.

Signed-off-by: Snorre Magnus Davøen <post@snorre.io>